JSR runs on eval_model code
